@@ -4,7 +4,7 @@ namespace SadSchool.ViewModels
 {
     public class ClassViewModel
     {
-        public int? Id { get; set; }
+        public int Id { get; set; }
         public string? Name { get; set; }
         public int? TeacherId { get; set; }
         public int? LeaderId { get; set; }
