@@ -20,7 +20,7 @@
                     {
                         <div class="col-4 d-flex justify-content-center">
                             <a class="btn btn-sad btn-main-sad blinker-sad" role="button"
-                               asp-controller="ClassBooks" asp-action="ClassBook" asp-route-className="@Model[i * 3 + j]">@Model[i * 3 + j]</a>
+                               asp-controller="ClassBooks" asp-action="ClassSelector" asp-route-className="@Model[i * 3 + j]">@Model[i * 3 + j]</a>
                         </div>
                     }
                 </div>
